Added note about filexch_ssh.pl bug and workaround.
There appears to be a bug in the filexch_ssh.pl script that can result in an inaccurate value appearing in the number of files transferred in the log file. This inaccurate value is used this script and so will appear in the output, including the summary table. 

The problem appears to arise when there is an empty string in the field corresponding to the <FilesBeginWith> key in a FileExchange .plist file. 

Workaround: never have an empty value for <FilesBeginWith>.